// packages/iam/ui/src/auth-ui/components/settings/shared/settings-card-types.ts
import type { UserAvatarClassNames } from "../../user-avatar";

export type SettingsCardClassNames = {
  base?: string;
  avatar?: UserAvatarClassNames;
  button?: string;
  cell?: string;
  checkbox?: string;
  destructiveButton?: string;
  content?: string;
  description?: string;
  dialog?: {
    content?: string;
    footer?: string;
    header?: string;
  };
  error?: string;
  footer?: string;
  header?: string;
  icon?: string;
  input?: string;
  instructions?: string;
  label?: string;
  primaryButton?: string;
  secondaryButton?: string;
  outlineButton?: string;
  skeleton?: string;
  title?: string;
};